+++ examples/egroup/main.tf
@@ -2,7 +2,7 @@ terraform {
   required_providers {
     cern = {
       source = "gitlab.cern.ch/batch-team/cern"
-      version = "1.0.0"
+      version = "> 1.0.0"
     }
   }
 }
@@ -13,4 +13,13 @@ data "cern_egroup" "batch" {
 
 output "batch_members" {
     value = data.cern_egroup.batch.members
+}
+
+data "cern_egroup" "ops" {
+    name        = "batch-operations"
+    query_mails = true
+}
+
+output "ops_mails" {
+    value = data.cern_egroup.ops.mails
 }